requirements.txt
aiohttp==3.6.2
aiomysql==0.0.20
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
google-api-core==1.15.0
google-auth==1.10.0
google-cloud-core==1.1.0
google-cloud-logging==1.14.0
googleapis-common-protos==1.6.0
grpcio==1.26.0
idna==2.8
iso8601==0.1.12
multidict==4.7.2
peewee==3.13.1
peewee-async==0.6.3a0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.2
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
urllib3==1.25.7
yarl==1.4.2
